fix: 64bit整数リテラルが32bitに切り詰められるバグを修正
- lexer.hpp: Token::Numberの型をintからlong longに変更し、atoiをatollに変更
- lexer.hpp: TokenStream::getCurNumVal()の戻り値をintからlong longに変更
- lexer.hpp: 重複していたgetCurNumVal()を削除、消えていたgetCurIndex/getCurTypeを復元
- codegen.cpp: generateNumber()がINT32範囲外の値をi64定数として生成するよう修正
- parser.cpp: TOK_DIGIT処理でint valとして受けていた箇所をlong longに修正

2000000000を超えるような大きな整数リテラルが、途中でint(32bit)を
経由することでオーバーフローし、意図しない負の値になっていた。
lexer/parser/codegenの複数箇所にintの受け取りが残っていたため、
すべてlong longに統一して修正した。

+++ parser.cpp
@@ -1192,7 +1192,7 @@ MemberArrayAccessAST *marr = new MemberArrayAccessAST(var_name, member_name, ind
     return new MemberAccessAST("this", member_name);
   }
   else if(Tokens->getCurType() == TOK_DIGIT){
-    int val = Tokens->getCurNumVal();
+    long long val = Tokens->getCurNumVal();
     Tokens->getNextToken();
     return new NumberAST(val);
   }
